+++ utils/src/Vulkan/Utils/ShaderQQ.hs
@@ -55,7 +55,7 @@ import           Vulkan.Utils.ShaderQQ.Interpolate
 -- An explicit example (@<interactive>@ is from doctest):
 --
 -- >>> let version = 450 :: Int in [glsl|#version $version|]
--- "#version 450\n#extension GL_GOOGLE_cpp_style_line_directive : enable\n#line 32 \"<interactive>\"\n"
+-- "#version 450\n#extension GL_GOOGLE_cpp_style_line_directive : enable\n#line 46 \"<interactive>\"\n"
 --
 -- Note that line number will be thrown off if any of the interpolated
 -- variables contain newlines.