+++ utils/changelog.md
@@ -6,6 +6,7 @@
 - Add `checkCommandsExp` function to generate an expression checking specified
   commands for non-nullness
 - Expose Queue family index for queues assigned with `assignQueues`
+- Add `Vulkan.Utils.ShaderQQ.Shaderc` to compile HLSL shaders
 
 ## [0.2] - 2020-11-15